Cleaning up tests for DTK classic adapters to only use unsigned long int as the global ordinal type

+++ packages/Adapters/ClassicDTK/test/geometry/tstGeometryManager.cpp
@@ -64,7 +64,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_cylinder_test )
     // Build a series of random cylinders.
     int num_cylinders = 100;
     Teuchos::ArrayRCP<Cylinder> cylinders( num_cylinders );
-    Teuchos::ArrayRCP<int> gids( num_cylinders );
+    Teuchos::ArrayRCP<unsigned long int> gids( num_cylinders );
     for ( int i = 0; i < num_cylinders; ++i )
     {
 	double length = (double) std::rand() / RAND_MAX;
@@ -78,7 +78,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_cylinder_test )
     }
 
     // Build a geometry manager.
-    GeometryManager<Cylinder,int> geometry_manager( cylinders, gids, comm, 3 );
+    GeometryManager<Cylinder,unsigned long int> geometry_manager( cylinders, gids, comm, 3 );
 
     // Check the geometry manager.
     Teuchos::ArrayRCP<Cylinder> manager_geometry = geometry_manager.geometry();
@@ -100,7 +100,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_cylinder_test )
 
     Teuchos::Array<BoundingBox> cylinder_boxes = 
 	geometry_manager.boundingBoxes();
-    Teuchos::ArrayRCP<int> cylinder_gids = geometry_manager.gids();
+    Teuchos::ArrayRCP<unsigned long int> cylinder_gids = geometry_manager.gids();
     for ( int i = 0; i < num_cylinders; ++i )
     {
 	TEST_ASSERT( cylinder_gids[i] == gids[i] );
@@ -160,7 +160,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_box_test )
     // Build a series of random boxes.
     int num_boxes = 100;
     Teuchos::ArrayRCP<Box> boxes( num_boxes );
-    Teuchos::ArrayRCP<int> gids( num_boxes );
+    Teuchos::ArrayRCP<unsigned long int> gids( num_boxes );
     for ( int i = 0; i < num_boxes; ++i )
     {
 	double x_min = -(double) std::rand() / RAND_MAX + my_rank;
@@ -175,7 +175,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_box_test )
     }
 
     // Build a geometry manager.
-    GeometryManager<Box,int> geometry_manager( boxes, gids, comm, 3 );
+    GeometryManager<Box,unsigned long int> geometry_manager( boxes, gids, comm, 3 );
 
     // Check the geometry manager.
     Teuchos::ArrayRCP<Box> manager_geometry = geometry_manager.geometry();
@@ -196,7 +196,7 @@ TEUCHOS_UNIT_TEST( GeometryManager, geometry_manager_box_test )
 
     Teuchos::Array<BoundingBox> box_boxes = 
 	geometry_manager.boundingBoxes();
-    Teuchos::ArrayRCP<int> box_gids = geometry_manager.gids();
+    Teuchos::ArrayRCP<unsigned long int> box_gids = geometry_manager.gids();
     for ( int i = 0; i < num_boxes; ++i )
     {
 	TEST_ASSERT( box_gids[i] == gids[i] );
