+++ packages/Adapters/ClassicDTK/test/map/tstSharedDomainMap10.cpp
@@ -142,11 +142,11 @@ class FieldTraits<MyField>
 //---------------------------------------------------------------------------//
 // FieldEvaluator Implementation.
 class MyEvaluator : 
-    public DataTransferKit::FieldEvaluator<int,MyField>
+    public DataTransferKit::FieldEvaluator<unsigned long int,MyField>
 {
   public:
 
-    MyEvaluator( const DataTransferKit::MeshContainer<int>& mesh, 
+    MyEvaluator( const DataTransferKit::MeshContainer<unsigned long int>& mesh, 
 		 const Teuchos::RCP< const Teuchos::Comm<int> >& comm )
 	: d_mesh( mesh )
 	, d_comm( comm )
@@ -157,7 +157,7 @@ class MyEvaluator :
 
     MyField evaluate( 
 	const Teuchos::ArrayRCP<
-	DataTransferKit::MeshContainer<int>::global_ordinal_type>& elements,
+	DataTransferKit::MeshContainer<unsigned long int>::global_ordinal_type>& elements,
 	const Teuchos::ArrayRCP<double>& coords )
     {
 	MyField evaluated_data( elements.size(), 1 );
@@ -179,20 +179,20 @@ class MyEvaluator :
 
   private:
 
-    DataTransferKit::MeshContainer<int>  d_mesh;
+    DataTransferKit::MeshContainer<unsigned long int>  d_mesh;
     Teuchos::RCP< const Teuchos::Comm<int> > d_comm;
 };
 
 //---------------------------------------------------------------------------//
 // Mesh create functions.
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> >
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> >
 buildTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 {
     // Make some vertices.
     int num_vertices = edge_length*edge_length;
     int vertex_dim = 2;
-    Teuchos::ArrayRCP<int> vertex_handles( num_vertices );
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles( num_vertices );
     Teuchos::ArrayRCP<double> coords( vertex_dim*num_vertices );
     int idx;
     for ( int j = 0; j < edge_length; ++j )
@@ -200,7 +200,7 @@ buildTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 	for ( int i = 0; i < edge_length; ++i )
 	{
 	    idx = i + j*edge_length;
-	    vertex_handles[ idx ] = (int) num_vertices*my_rank + idx + elem_offset;
+	    vertex_handles[ idx ] = (unsigned long int) num_vertices*my_rank + idx + elem_offset;
 	    coords[ idx ] = i + my_rank*(edge_length-1);
 	    coords[ num_vertices + idx ] = j;
 	}
@@ -208,9 +208,9 @@ buildTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     
     // Make the triangles. 
     int num_elements = (edge_length-1)*(edge_length-1)*2;
-    Teuchos::ArrayRCP<int> tri_handles( num_elements );
-    Teuchos::ArrayRCP<int> tri_connectivity( 3*num_elements );
-    int elem_idx, vertex_idx;
+    Teuchos::ArrayRCP<unsigned long int> tri_handles( num_elements );
+    Teuchos::ArrayRCP<unsigned long int> tri_connectivity( 3*num_elements );
+    unsigned long int elem_idx, vertex_idx;
     int v0, v1, v2, v3;
     for ( int j = 0; j < (edge_length-1); ++j )
     {
@@ -246,20 +246,20 @@ buildTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     }
 
     return Teuchos::rcp( 
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_TRIANGLE, 3,
-						 tri_handles, tri_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_TRIANGLE, 3,
+							       tri_handles, tri_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> >
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> >
 buildTiledTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 {
     // Make some vertices.
     int num_vertices = edge_length*edge_length;
     int vertex_dim = 2;
-    Teuchos::ArrayRCP<int> vertex_handles( num_vertices );
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles( num_vertices );
     Teuchos::ArrayRCP<double> coords( vertex_dim*num_vertices );
     int idx;
     for ( int j = 0; j < edge_length; ++j )
@@ -267,7 +267,7 @@ buildTiledTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 	for ( int i = 0; i < edge_length; ++i )
 	{
 	    idx = i + j*edge_length;
-	    vertex_handles[ idx ] = (int) num_vertices*my_rank + idx + elem_offset;
+	    vertex_handles[ idx ] = (unsigned long int) num_vertices*my_rank + idx + elem_offset;
 	    coords[ idx ] = i + my_rank*(edge_length-1);
 	    coords[ num_vertices + idx ] = j + my_rank*(edge_length-1);
 	}
@@ -275,9 +275,9 @@ buildTiledTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     
     // Make the triangles. 
     int num_elements = (edge_length-1)*(edge_length-1)*2;
-    Teuchos::ArrayRCP<int> tri_handles( num_elements );
-    Teuchos::ArrayRCP<int> tri_connectivity( 3*num_elements );
-    int elem_idx, vertex_idx;
+    Teuchos::ArrayRCP<unsigned long int> tri_handles( num_elements );
+    Teuchos::ArrayRCP<unsigned long int> tri_connectivity( 3*num_elements );
+    unsigned long int elem_idx, vertex_idx;
     int v0, v1, v2, v3;
     for ( int j = 0; j < (edge_length-1); ++j )
     {
@@ -313,19 +313,19 @@ buildTiledTriMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     }
 
     return Teuchos::rcp(
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_TRIANGLE, 3,
-						 tri_handles, tri_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_TRIANGLE, 3,
+							       tri_handles, tri_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> > buildNullTriMesh()
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> > buildNullTriMesh()
 {
-    Teuchos::ArrayRCP<int> vertex_handles(0);
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles(0);
     Teuchos::ArrayRCP<double> coords(0);
-    Teuchos::ArrayRCP<int> tri_handles(0);
-    Teuchos::ArrayRCP<int> tri_connectivity(0);
+    Teuchos::ArrayRCP<unsigned long int> tri_handles(0);
+    Teuchos::ArrayRCP<unsigned long int> tri_connectivity(0);
     Teuchos::ArrayRCP<int> permutation_list(3);
     for ( int i = 0; (int) i < permutation_list.size(); ++i )
     {
@@ -333,20 +333,20 @@ Teuchos::RCP<DataTransferKit::MeshContainer<int> > buildNullTriMesh()
     }
 
     return Teuchos::rcp( 
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_TRIANGLE, 3,
-						 tri_handles, tri_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_TRIANGLE, 3,
+							       tri_handles, tri_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> >  
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> >  
 buildQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 {
     // Make some vertices.
     int num_vertices = edge_length*edge_length;
     int vertex_dim = 2;
-    Teuchos::ArrayRCP<int> vertex_handles( num_vertices );
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles( num_vertices );
     Teuchos::ArrayRCP<double> coords( vertex_dim*num_vertices );
     int idx;
     for ( int j = 0; j < edge_length; ++j )
@@ -354,7 +354,7 @@ buildQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 	for ( int i = 0; i < edge_length; ++i )
 	{
 	    idx = i + j*edge_length;
-	    vertex_handles[ idx ] = (int) num_vertices*my_rank + idx + elem_offset;
+	    vertex_handles[ idx ] = (unsigned long int) num_vertices*my_rank + idx + elem_offset;
 	    coords[ idx ] = i + my_rank*(edge_length-1);
 	    coords[ num_vertices + idx ] = j;
 	}
@@ -362,8 +362,8 @@ buildQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     
     // Make the quadrilaterals. 
     int num_elements = (edge_length-1)*(edge_length-1);
-    Teuchos::ArrayRCP<int> quad_handles( num_elements );
-    Teuchos::ArrayRCP<int> quad_connectivity( 4*num_elements );
+    Teuchos::ArrayRCP<unsigned long int> quad_handles( num_elements );
+    Teuchos::ArrayRCP<unsigned long int> quad_connectivity( 4*num_elements );
     int elem_idx, vertex_idx;
     for ( int j = 0; j < (edge_length-1); ++j )
     {
@@ -395,20 +395,20 @@ buildQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     }
 
     return Teuchos::rcp( 
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_QUADRILATERAL, 4,
-						 quad_handles, quad_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_QUADRILATERAL, 4,
+							       quad_handles, quad_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> >  
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> >  
 buildTiledQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 {
     // Make some vertices.
     int num_vertices = edge_length*edge_length;
     int vertex_dim = 2;
-    Teuchos::ArrayRCP<int> vertex_handles( num_vertices );
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles( num_vertices );
     Teuchos::ArrayRCP<double> coords( vertex_dim*num_vertices );
     int idx;
     for ( int j = 0; j < edge_length; ++j )
@@ -416,7 +416,7 @@ buildTiledQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
 	for ( int i = 0; i < edge_length; ++i )
 	{
 	    idx = i + j*edge_length;
-	    vertex_handles[ idx ] = (int) num_vertices*my_rank + idx + elem_offset;
+	    vertex_handles[ idx ] = (unsigned long int) num_vertices*my_rank + idx + elem_offset;
 	    coords[ idx ] = i + my_rank*(edge_length-1);
 	    coords[ num_vertices + idx ] = j + my_rank*(edge_length-1);
 	}
@@ -424,8 +424,8 @@ buildTiledQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     
     // Make the quadrilaterals. 
     int num_elements = (edge_length-1)*(edge_length-1);
-    Teuchos::ArrayRCP<int> quad_handles( num_elements );
-    Teuchos::ArrayRCP<int> quad_connectivity( 4*num_elements );
+    Teuchos::ArrayRCP<unsigned long int> quad_handles( num_elements );
+    Teuchos::ArrayRCP<unsigned long int> quad_connectivity( 4*num_elements );
     int elem_idx, vertex_idx;
     for ( int j = 0; j < (edge_length-1); ++j )
     {
@@ -457,20 +457,20 @@ buildTiledQuadMesh( int my_rank, int my_size, int edge_length, int elem_offset )
     }
 
     return Teuchos::rcp( 
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_QUADRILATERAL, 4,
-						 quad_handles, quad_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_QUADRILATERAL, 4,
+							       quad_handles, quad_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
-Teuchos::RCP<DataTransferKit::MeshContainer<int> > 
+Teuchos::RCP<DataTransferKit::MeshContainer<unsigned long int> > 
 buildNullQuadMesh()
 {
-    Teuchos::ArrayRCP<int> vertex_handles(0);
+    Teuchos::ArrayRCP<unsigned long int> vertex_handles(0);
     Teuchos::ArrayRCP<double> coords(0);
-    Teuchos::ArrayRCP<int> quad_handles(0);
-    Teuchos::ArrayRCP<int> quad_connectivity(0);
+    Teuchos::ArrayRCP<unsigned long int> quad_handles(0);
+    Teuchos::ArrayRCP<unsigned long int> quad_connectivity(0);
     Teuchos::ArrayRCP<int> permutation_list(4);
     for ( int i = 0; (int) i < permutation_list.size(); ++i )
     {
@@ -478,10 +478,10 @@ buildNullQuadMesh()
     }
 
     return Teuchos::rcp( 
-	new DataTransferKit::MeshContainer<int>( 2, vertex_handles, coords, 
-						 DataTransferKit::DTK_QUADRILATERAL, 4,
-						 quad_handles, quad_connectivity,
-						 permutation_list ) );
+	new DataTransferKit::MeshContainer<unsigned long int>( 2, vertex_handles, coords, 
+							       DataTransferKit::DTK_QUADRILATERAL, 4,
+							       quad_handles, quad_connectivity,
+							       permutation_list ) );
 }
 
 //---------------------------------------------------------------------------//
@@ -546,7 +546,7 @@ int num_points = 1000;
 TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_test10 )
 {
     using namespace DataTransferKit;
-    typedef MeshContainer<int> MeshType;
+    typedef MeshContainer<unsigned long int> MeshType;
 
     // Setup communication.
     Teuchos::RCP< const Teuchos::Comm<int> > comm = getDefaultComm<int>();
@@ -602,7 +602,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_test10 )
 	Teuchos::rcp( new FieldManager<MyField>( coordinate_field, comm ) );
 
     // Create field evaluator.
-    Teuchos::RCP< FieldEvaluator<int,MyField> > source_evaluator;
+    Teuchos::RCP< FieldEvaluator<unsigned long int,MyField> > source_evaluator;
     if ( my_rank == 0 )
     {
     	source_evaluator = Teuchos::rcp( new MyEvaluator( *mesh_blocks[0], comm ) );
@@ -655,7 +655,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_test10 )
 TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_expanded_test10 )
 {
     using namespace DataTransferKit;
-    typedef MeshContainer<int> MeshType;
+    typedef MeshContainer<unsigned long int> MeshType;
 
     // Setup communication.
     Teuchos::RCP< const Teuchos::Comm<int> > comm = getDefaultComm<int>();
@@ -711,7 +711,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_expanded_test10 )
 	Teuchos::rcp( new FieldManager<MyField>( coordinate_field, comm ) );
 
     // Create field evaluator.
-    Teuchos::RCP<FieldEvaluator<int,MyField> > source_evaluator;
+    Teuchos::RCP<FieldEvaluator<unsigned long int,MyField> > source_evaluator;
     if ( my_rank == 0 )
     {
     	source_evaluator = Teuchos::rcp( new MyEvaluator( *mesh_blocks[0], comm ) );
@@ -746,7 +746,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_expanded_test10 )
     // Check the data transfer. Each target point should have been assigned
     // its source rank + 1 as data if it is in the mesh and 0.0 if it is outside.
     double source_rank;
-    Teuchos::Array<int> missing_points;
+    Teuchos::Array<unsigned long int> missing_points;
     for ( int n = 0; n < num_points; ++n )
     {
 	if ( *(coordinate_field->begin()+n) < 0.0 ||
@@ -779,7 +779,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_expanded_test10 )
 
     // Check the missing points.
     TEST_ASSERT( missing_points.size() > 0 );
-    Teuchos::ArrayView<int> missed_in_map = 
+    Teuchos::ArrayView<unsigned long int> missed_in_map = 
 	shared_domain_map.getMissedTargetPoints();
     TEST_EQUALITY( missing_points.size(), missed_in_map.size() );
 
@@ -799,7 +799,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_expanded_test10 )
 TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_tiled_test10 )
 {
     using namespace DataTransferKit;
-    typedef MeshContainer<int> MeshType;
+    typedef MeshContainer<unsigned long int> MeshType;
 
     // Setup communication.
     Teuchos::RCP< const Teuchos::Comm<int> > comm = getDefaultComm<int>();
@@ -855,7 +855,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_tiled_test10 )
 	Teuchos::rcp( new FieldManager<MyField>( coordinate_field, comm ) );
 
     // Create field evaluator.
-    Teuchos::RCP<FieldEvaluator<int,MyField> > source_evaluator;
+    Teuchos::RCP<FieldEvaluator<unsigned long int,MyField> > source_evaluator;
     if ( my_rank == 0 )
     {
     	source_evaluator = Teuchos::rcp( new MyEvaluator( *mesh_blocks[0], comm ) );
@@ -890,7 +890,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_tiled_test10 )
     // Check the data transfer. Each target point should have been assigned
     // its source rank + 1 as data if it is in the mesh and 0.0 if it is outside.
     double source_rank;
-    Teuchos::Array<int> missing_points;
+    Teuchos::Array<unsigned long int> missing_points;
     bool tagged;
     for ( int n = 0; n < num_points; ++n )
     {
@@ -940,7 +940,7 @@ TEUCHOS_UNIT_TEST( SharedDomainMap, shared_domain_map_tiled_test10 )
 
     // Check the missing points.
     TEST_ASSERT( missing_points.size() > 0 );
-    Teuchos::ArrayView<int> missed_in_map = 
+    Teuchos::ArrayView<unsigned long int> missed_in_map = 
 	shared_domain_map.getMissedTargetPoints();
     TEST_EQUALITY( missing_points.size(), missed_in_map.size() );
